Start to replace Int with BigInt

+++ compiler/lib/src/main/scala/analysis/Semantics/Command.scala
@@ -16,7 +16,7 @@ sealed trait Command {
 
 object Command {
 
-  type Opcode = Int
+  type Opcode = BigInt
 
   /** A non-parameter command */
   final case class NonParam(
